refactor(home): scope useHomeDnd to drag-and-drop only
The hook no longer loads data nor returns hasLoaded/isAppsLoading/apps;
the component reads those from useHomeLayout and passes items, layout and
saveLayout in. useHomeDnd now returns only DnD state.
